Added --keep-agent

+++ particles-optimiser/production/cleanup.sh
@@ -4,4 +4,6 @@ kubectl delete -f particles-v1.0.0.yaml
 kubectl delete -f particles-v1.0.1.yaml
 kubectl delete -f particles-v1.0.2.yaml
 
-kubectl delete ns production
+if [ "$1" != "--keep-agent" ]; then
+    kubectl delete ns production
+fi
